+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+3.0.0
+-----
+
+* The library now exports an ESM-compatible arucoToSVGString function rather than an object-based API.
+* Added an HTML custom element `<aruco-marker markerid="155"></aruco-marker>` that works regardless of framework.
+* Removed AngularJS component.
+* Removed AMD and Bower support.
+
 2.0.0
 -----